Extract optimizer.
