+++ autoprecompiles/src/optimizer.rs
@@ -1,0 +1,143 @@
+use std::collections::BTreeMap;
+
+use powdr_ast::analyzed::AlgebraicExpression;
+use powdr_constraint_solver::constraint_system::BusInteractionHandler;
+use powdr_number::FieldElement;
+use powdr_pilopt::simplify_expression;
+
+use crate::{
+    constraint_optimizer::{optimize_constraints, IsBusStateful},
+    powdr,
+    register_optimizer::{check_register_operation_consistency, optimize_register_operations},
+    SymbolicMachine, EXECUTION_BUS_ID, PC_LOOKUP_BUS_ID,
+};
+
+pub fn optimize<T: FieldElement>(
+    machine: SymbolicMachine<T>,
+    bus_interaction_handler: impl BusInteractionHandler<T> + IsBusStateful<T> + 'static + Clone,
+    opcode: u32,
+    degree_bound: usize,
+) -> SymbolicMachine<T> {
+    let machine = optimize_pc_lookup(machine, opcode);
+    let machine = optimize_exec_bus(machine);
+    assert!(check_register_operation_consistency(&machine));
+
+    // We need to remove memory bus interactions with inlined multiplicity zero before
+    // doing register memory optimizations.
+    let machine = optimize_constraints(machine, bus_interaction_handler.clone(), degree_bound);
+    assert!(check_register_operation_consistency(&machine));
+
+    let machine = optimize_register_operations(machine);
+    assert!(check_register_operation_consistency(&machine));
+
+    // Fixpoint style re-attempt.
+    // TODO we probably need proper fixpoint here at some point.
+    let machine = optimize_constraints(machine, bus_interaction_handler, degree_bound);
+    assert!(check_register_operation_consistency(&machine));
+    machine
+}
+
+pub fn optimize_pc_lookup<T: FieldElement>(
+    mut machine: SymbolicMachine<T>,
+    opcode: u32,
+) -> SymbolicMachine<T> {
+    let mut first_pc = None;
+    machine.bus_interactions.retain(|bus_int| {
+        if bus_int.id == PC_LOOKUP_BUS_ID {
+            if first_pc.is_none() {
+                first_pc = Some(bus_int.clone());
+            }
+            return false;
+        }
+        true
+    });
+    let mut first_pc = first_pc.unwrap();
+    assert_eq!(first_pc.args.len(), 9);
+    first_pc.args[1] = AlgebraicExpression::Number(T::from(opcode));
+    first_pc.args[2] = AlgebraicExpression::Number(T::from(0u32));
+    first_pc.args[3] = AlgebraicExpression::Number(T::from(0u32));
+    first_pc.args[4] = AlgebraicExpression::Number(T::from(0u32));
+    first_pc.args[5] = AlgebraicExpression::Number(T::from(0u32));
+    first_pc.args[6] = AlgebraicExpression::Number(T::from(0u32));
+    first_pc.args[7] = AlgebraicExpression::Number(T::from(0u32));
+    first_pc.args[8] = AlgebraicExpression::Number(T::from(0u32));
+
+    machine.bus_interactions.push(first_pc);
+
+    machine
+}
+
+pub fn optimize_exec_bus<T: FieldElement>(mut machine: SymbolicMachine<T>) -> SymbolicMachine<T> {
+    let mut first_seen = false;
+    let mut receive = true;
+    let mut latest_send = None;
+    let mut subs_pc: BTreeMap<AlgebraicExpression<T>, AlgebraicExpression<T>> = Default::default();
+    let mut subs_ts: BTreeMap<AlgebraicExpression<T>, AlgebraicExpression<T>> = Default::default();
+    machine.bus_interactions.retain(|bus_int| {
+        if bus_int.id != EXECUTION_BUS_ID {
+            return true;
+        }
+
+        if receive {
+            // TODO assert that mult matches -expr
+        }
+
+        // Keep the first receive
+        let keep = if !first_seen {
+            first_seen = true;
+            true
+        } else if !receive {
+            // Save the latest send and remove the bus interaction
+            let mut pc_expr = bus_int.args[0].clone();
+            powdr::substitute_algebraic_algebraic(&mut pc_expr, &subs_pc);
+            pc_expr = simplify_expression(pc_expr);
+
+            let mut ts_expr = bus_int.args[1].clone();
+            powdr::substitute_algebraic_algebraic(&mut ts_expr, &subs_ts);
+            ts_expr = simplify_expression(ts_expr);
+
+            let mut send = bus_int.clone();
+            send.args[0] = pc_expr;
+            send.args[1] = ts_expr;
+
+            latest_send = Some(send);
+            false
+        } else {
+            // Equate the latest send to the new receive and remove the bus interaction
+            subs_pc.insert(
+                bus_int.args[0].clone(),
+                latest_send.clone().unwrap().args[0].clone(),
+            );
+            subs_ts.insert(
+                bus_int.args[1].clone(),
+                latest_send.clone().unwrap().args[1].clone(),
+            );
+            false
+        };
+
+        receive = !receive;
+
+        keep
+    });
+
+    // Re-add the last send
+    machine.bus_interactions.push(latest_send.unwrap());
+
+    for c in &mut machine.constraints {
+        powdr::substitute_algebraic_algebraic(&mut c.expr, &subs_pc);
+        powdr::substitute_algebraic_algebraic(&mut c.expr, &subs_ts);
+        c.expr = simplify_expression(c.expr.clone());
+    }
+    for b in &mut machine.bus_interactions {
+        powdr::substitute_algebraic_algebraic(&mut b.mult, &subs_pc);
+        powdr::substitute_algebraic_algebraic(&mut b.mult, &subs_ts);
+        b.mult = simplify_expression(b.mult.clone());
+        for a in &mut b.args {
+            powdr::substitute_algebraic_algebraic(a, &subs_pc);
+            powdr::substitute_algebraic_algebraic(a, &subs_ts);
+            *a = simplify_expression(a.clone());
+        }
+    }
+
+    machine
+}
